# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgroGestion API",
    "description": "API para gestionar cultivos de principio a fin: siembras, ciclos, conteo de plantas, vivero, riesgos y eventos adversos. Todos los errores usan el mismo formato: decida siempre con el campo `error`, nunca con `message`.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/cultivos": {
      "get": {
        "tags": [
          "4. Cultivos"
        ],
        "summary": "Buscar cultivos",
        "operationId": "listar_cultivos_api_v1_cultivos_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 80
                },
                {
                  "type": "null"
                }
              ],
              "description": "Texto a buscar en el nombre",
              "title": "Q"
            },
            "description": "Texto a buscar en el nombre"
          },
          {
            "name": "tipo_ciclo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "transitorio",
                    "semipermanente",
                    "permanente",
                    "forestal"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipo Ciclo"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Registros a saltar",
              "default": 0,
              "title": "Skip"
            },
            "description": "Registros a saltar"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Tamaño de página, máximo 50",
              "default": 10,
              "title": "Limit"
            },
            "description": "Tamaño de página, máximo 50"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CultivoResumen_"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "4. Cultivos"
        ],
        "summary": "Crear un cultivo con su perfil",
        "operationId": "crear_cultivo_api_v1_cultivos_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CultivoCrear"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CultivoSalida"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cultivos/{cultivo_id}": {
      "get": {
        "tags": [
          "4. Cultivos"
        ],
        "summary": "Ver el perfil de un cultivo",
        "operationId": "ver_cultivo_api_v1_cultivos__cultivo_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cultivo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cultivo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CultivoSalida"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "4. Cultivos"
        ],
        "summary": "Reemplazar el perfil de un cultivo",
        "operationId": "actualizar_cultivo_api_v1_cultivos__cultivo_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cultivo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cultivo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CultivoActualizar"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CultivoSalida"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cultivos/{cultivo_id}/fases": {
      "get": {
        "tags": [
          "4. Cultivos"
        ],
        "summary": "Fases del cultivo",
        "operationId": "fases_api_v1_cultivos__cultivo_id__fases_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cultivo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cultivo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FaseSalida"
                  },
                  "title": "Response Fases Api V1 Cultivos  Cultivo Id  Fases Get"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cultivos/{cultivo_id}/metodos": {
      "get": {
        "tags": [
          "4. Cultivos"
        ],
        "summary": "Métodos de propagación",
        "operationId": "metodos_api_v1_cultivos__cultivo_id__metodos_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cultivo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cultivo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MetodoSalida"
                  },
                  "title": "Response Metodos Api V1 Cultivos  Cultivo Id  Metodos Get"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cultivos/{cultivo_id}/dosis": {
      "get": {
        "tags": [
          "4. Cultivos"
        ],
        "summary": "Dosis de referencia",
        "operationId": "dosis_api_v1_cultivos__cultivo_id__dosis_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cultivo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cultivo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DosisSalida"
                  },
                  "title": "Response Dosis Api V1 Cultivos  Cultivo Id  Dosis Get"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cultivos/{cultivo_id}/riesgos": {
      "get": {
        "tags": [
          "4. Cultivos"
        ],
        "summary": "Riesgos del cultivo",
        "operationId": "riesgos_cultivo_api_v1_cultivos__cultivo_id__riesgos_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cultivo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cultivo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CultivoRiesgoSalida"
                  },
                  "title": "Response Riesgos Cultivo Api V1 Cultivos  Cultivo Id  Riesgos Get"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "4. Cultivos"
        ],
        "summary": "Reemplazar los riesgos del cultivo",
        "operationId": "reemplazar_riesgos_cultivo_api_v1_cultivos__cultivo_id__riesgos_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cultivo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cultivo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CultivoRiesgoEntrada"
                },
                "title": "Datos"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CultivoRiesgoSalida"
                  },
                  "title": "Response Reemplazar Riesgos Cultivo Api V1 Cultivos  Cultivo Id  Riesgos Put"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CultivoSalida": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "grupo": {
            "type": "string",
            "title": "Grupo"
          },
          "tipo_ciclo": {
            "type": "string",
            "title": "Tipo Ciclo"
          },
          "unidad_conteo": {
            "type": "string",
            "title": "Unidad Conteo"
          },
          "por_validar": {
            "type": "boolean",
            "title": "Por Validar"
          },
          "nombre_cientifico": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre Cientifico"
          },
          "tipo_renovacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tipo Renovacion"
          },
          "unidad_cosecha": {
            "type": "string",
            "title": "Unidad Cosecha"
          },
          "pago_cosecha_sugerido": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pago Cosecha Sugerido"
          },
          "densidad_ref": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Densidad Ref"
          },
          "meses_primera_cosecha": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meses Primera Cosecha"
          },
          "cosechas_por_anio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cosechas Por Anio"
          },
          "fuente": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fuente"
          },
          "fases": {
            "items": {
              "$ref": "#/components/schemas/FaseSalida"
            },
            "type": "array",
            "title": "Fases"
          },
          "metodos": {
            "items": {
              "$ref": "#/components/schemas/MetodoSalida"
            },
            "type": "array",
            "title": "Metodos"
          },
          "dosis": {
            "items": {
              "$ref": "#/components/schemas/DosisSalida"
            },
            "type": "array",
            "title": "Dosis"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre",
          "grupo",
          "tipo_ciclo",
          "unidad_conteo",
          "por_validar",
          "nombre_cientifico",
          "tipo_renovacion",
          "unidad_cosecha",
          "pago_cosecha_sugerido",
          "densidad_ref",
          "meses_primera_cosecha",
          "cosechas_por_anio",
          "fuente",
          "fases",
          "metodos",
          "dosis"
        ],
        "title": "CultivoSalida"
      },
      "DetalleError": {
        "properties": {
          "campo": {
            "type": "string",
            "title": "Campo"
          },
          "mensaje": {
            "type": "string",
            "title": "Mensaje"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          }
        },
        "type": "object",
        "required": [
          "campo",
          "mensaje",
          "tipo"
        ],
        "title": "DetalleError"
      },
...
      "ErrorRespuesta": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DetalleError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message",
          "status_code"
        ],
        "title": "ErrorRespuesta",
        "description": "Formato único de error: decida siempre con `error`, nunca con `message`."
      },
...
      "Page_CultivoResumen_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CultivoResumen"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "has_more"
        ],
        "title": "Page[CultivoResumen]"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}